feat: helper function to constrain column type
