+++ metric.go
@@ -13,194 +13,6 @@ import (
 	"github.com/apache/arrow/go/arrow/flight"
 )
 
-type column struct {
-	typ      greptimepb.ColumnDataType
-	semantic greptimepb.Column_SemanticType
-}
-
-func checkColumnEquality(key string, col1, col2 column) error {
-	if col1.typ != col2.typ {
-		return fmt.Errorf("the type of '%s' does not match: '%T' and '%T'", key, col1.typ, col2.typ)
-	}
-	if col1.semantic != col2.semantic {
-		return fmt.Errorf("tag and field MUST NOT contain same name")
-	}
-
-	return nil
-}
-
-// Series represents one row of data you want to insert into GreptimeDB.
-//   - Tag fields are the index columns, which helps you to query data efficiently
-//   - Field fields are the value columns, which are used for value
-//   - Timestamp field is the timestamp column, which is required
-//
-// you do not need to create schema in advance, it will be created based on Series.
-// But once the schema is created, [Client] has no ability to alert it.
-type Series struct {
-	orders  []string
-	columns map[string]column
-	vals    map[string]any
-
-	timestamp time.Time // required
-}
-
-// GetTagsAndFields get all column names from metric, except timestamp column
-func (s *Series) GetTagsAndFields() []string {
-	dst := make([]string, len(s.orders))
-	copy(dst, s.orders)
-	return dst
-}
-
-// Get helps to get value of specifid column. The second return value
-// indicates if the key was present in Series
-func (s *Series) Get(key string) (any, bool) {
-	val, exist := s.vals[key]
-	return val, exist
-}
-
-func (s *Series) GetUint(key string) (uint64, bool) {
-	val, exist := s.Get(key)
-	if !exist {
-		return 0, exist
-	}
-
-	switch val.(type) {
-	case uint64:
-		return val.(uint64), true
-	case uint32:
-		return uint64(val.(uint32)), true
-	case uint16:
-		return uint64(val.(uint16)), true
-	case uint8:
-		return uint64(val.(uint8)), true
-	default:
-		return 0, false
-	}
-}
-
-func (s *Series) GetInt(key string) (int64, bool) {
-	val, exist := s.Get(key)
-	if !exist {
-		return 0, exist
-	}
-
-	switch val.(type) {
-	case int64:
-		return val.(int64), true
-	case int32:
-		return int64(val.(int32)), true
-	case int16:
-		return int64(val.(int16)), true
-	case int8:
-		return int64(val.(int8)), true
-	default:
-		return 0, false
-	}
-}
-
-func (s *Series) GetFloat(key string) (float64, bool) {
-	val, exist := s.Get(key)
-	if !exist {
-		return 0, exist
-	}
-
-	switch val.(type) {
-	case float64:
-		return val.(float64), true
-	case float32:
-		return float64(val.(float32)), true
-	default:
-		return 0, false
-	}
-}
-
-func (s *Series) GetBool(key string) (bool, bool) {
-	val, exist := s.Get(key)
-	if !exist {
-		return false, exist
-	}
-
-	v, ok := val.(bool)
-	return v, ok
-}
-
-func (s *Series) GetString(key string) (string, bool) {
-	val, exist := s.Get(key)
-	if !exist {
-		return "", exist
-	}
-
-	v, ok := val.(string)
-	return v, ok
-}
-
-func (s *Series) GetBytes(key string) ([]byte, bool) {
-	val, exist := s.GetString(key)
-	if !exist {
-		return nil, exist
-	}
-
-	return []byte(val), true
-}
-
-// GetTimestamp get timestamp field
-func (s *Series) GetTimestamp() time.Time {
-	return s.timestamp
-}
-
-func (s *Series) add(name string, val any, semantic greptimepb.Column_SemanticType) error {
-	key, err := toColumnName(name)
-	if err != nil {
-		return err
-	}
-
-	if s.columns == nil {
-		s.columns = map[string]column{}
-	}
-
-	v, err := convert(val)
-	if err != nil {
-		return fmt.Errorf("add tag err: %w", err)
-	}
-
-	newCol := column{
-		typ:      v.typ,
-		semantic: semantic,
-	}
-	if col, seen := s.columns[key]; seen {
-		if err := checkColumnEquality(key, col, newCol); err != nil {
-			return err
-		}
-	}
-	s.columns[key] = newCol
-	s.orders = append(s.orders, key)
-
-	if s.vals == nil {
-		s.vals = map[string]any{}
-	}
-	s.vals[key] = v.val
-
-	return nil
-}
-
-// AddTag prepare tag column, and old value will be replaced if same tag is set.
-// the length of key CAN NOT be longer than 100
-func (s *Series) AddTag(key string, val any) error {
-	return s.add(key, val, greptimepb.Column_TAG)
-}
-
-// AddField prepare field column, and old value will be replaced if same field is set.
-// the length of key CAN NOT be longer than 100
-func (s *Series) AddField(key string, val any) error {
-	return s.add(key, val, greptimepb.Column_FIELD)
-}
-
-// SetTimestamp is required
-func (s *Series) SetTimestamp(t time.Time) error {
-	s.timestamp = t
-	return nil
-}
-
 // Metric represents multiple rows of data, and also Metric can specify
 // the timestamp column name and precision
 type Metric struct {
